+++ CHANGELOG.md
@@ -9,6 +9,12 @@ This project adheres to Semantic Versioning.
 ### Added
 
 * `graphics::set_letterbox_color` allows you to set the color of the letterbox bars shown in certain scaling modes.
+* Basic support for off-screen rendering/'render to texture' has been implemented, in the form of the `Canvas` object.
+
+### Changed
+
+* Alpha blending should now work in a more predictable way. This may need further tweaks later on.
+* The renderer now flips drawing automatically when drawing to a framebuffer, due to how OpenGL stores textures. This is similar to how FNA and Love2D handle the same problem.
 
 ## [0.2.13] - 2019-03-05